requirements.txt
# Core scientific libraries
networkx>=2.8.0
numpy>=1.22.0
matplotlib>=3.5.0
scipy>=1.8.0
pyyaml>=6.0.0  # For configuration file handling
cryptography>=40.0.0  # For secure credential handling

# IBM Quantum (complete stack)
qiskit==1.4.1
qiskit-aer==0.16.4
qiskit-ibm-runtime==0.36.1
qiskit-ibm-provider==0.11.0
qiskit-qasm3-import==0.5.1  # OpenQASM 3.0 import support for Qiskit

# AWS Braket
amazon-braket-sdk==1.90.1
amazon-braket-schemas==1.23.1
amazon-braket-default-simulator==1.26.4
boto3>=1.24.0

# Google Quantum (complete stack)
cirq==1.4.1
cirq-core==1.4.1
cirq-google==1.4.1
cirq-aqt==1.4.1
cirq-ionq==1.4.1
cirq-pasqal==1.4.1
cirq-rigetti==1.4.1
cirq-web==1.4.1
protobuf>=3.20.0,<4.0.0  # Required by Cirq for compatibility

# OpenQASM support
openqasm3>=0.4.0  # General OpenQASM 3.0 parsing

# Cross-platform tools
pyquil>=3.0.0  # For Rigetti compatibility
qutip>=4.7.0   # Quantum toolbox in Python

# Development tools
jupyter>=1.0.0  # For interactive development

# Authentication and security
requests>=2.28.0  # For API interactions
python-dotenv>=1.0.0  # For loading environment variables
